mudando a logica do header
