+++ projeto-pokedex/.eslintcache
@@ -1,1 +1,1 @@
-
+[{"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\index.js":"1","C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\router\\Router.js":"2","C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\router\\coordinator.js":"3","C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\pages\\DetailsPage.js":"4","C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\pages\\ErrorPage.js":"5","C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\components\\Footer\\Footer.js":"6","C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\hooks\\useRequestData.js":"7","C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\components\\MainHome\\Buttons.js":"8","C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\components\\Header.js":"9","C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\components\\MainHome\\CardPokemon.js":"10"},{"size":450,"mtime":1606158065322,"results":"11","hashOfConfig":"12"},{"size":1450,"mtime":1606308663824,"results":"13","hashOfConfig":"12"},{"size":223,"mtime":1606220714827,"results":"14","hashOfConfig":"12"},{"size":174,"mtime":1606307593836,"results":"15","hashOfConfig":"12"},{"size":169,"mtime":1606307583216,"results":"16","hashOfConfig":"12"},{"size":512,"mtime":1606306787241,"results":"17","hashOfConfig":"12"},{"size":453,"mtime":1606306787318,"results":"18","hashOfConfig":"12"},{"size":468,"mtime":1606308867371,"results":"19","hashOfConfig":"12"},{"size":1139,"mtime":1606308632845,"results":"20","hashOfConfig":"12"},{"size":937,"mtime":1606308885092,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"9tgjk2",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\index.js",[],"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\router\\Router.js",[],"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\router\\coordinator.js",[],"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\pages\\DetailsPage.js",[],"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\pages\\ErrorPage.js",[],"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\components\\Footer\\Footer.js",[],"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\hooks\\useRequestData.js",[],"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\components\\MainHome\\Buttons.js",[],"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\components\\Header.js",[],"C:\\Users\\Nicole\\Desktop\\labenu\\dumont-pokedex4\\projeto-pokedex\\src\\components\\MainHome\\CardPokemon.js",[]]
