+++ projeto-pokedex/src/pages/DetailsPage.js
@@ -1,13 +1,9 @@
 import React from 'react'
-import Header from '../components/MainDetails/Header'
-import Footer from '../components/Footer/Footer'
 
 function DetailsPage() {
     return (
       <div>
-        <Header />
         <p>DetailsPage</p>
-        <Footer />
       </div>
     );
   }
