+++ projeto-pokedex/src/pages/ErrorPage.js
@@ -1,13 +1,9 @@
 import React from 'react'
-import Header from '../components/MainHome/Header'
-import Footer from '../components/Footer/Footer'
 
 function ErrorPage() {
     return (
       <div>
-          <Header />
         <p>nada aqui :(</p>
-        <Footer />
     </div>
     );
   }
